--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -1,12 +1,11 @@
 package hw10programoptimization
 
 import (
-	"encoding/json"
-	"fmt"
+	"bufio"
 	"io"
-	"io/ioutil"
-	"regexp"
 	"strings"
+
+	"github.com/mailru/easyjson"
 )
 
 type User struct {
@@ -21,47 +20,78 @@ type User struct {
 
 type DomainStat map[string]int
 
-func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
-	u, err := getUsers(r)
-	if err != nil {
-		return nil, fmt.Errorf("get users error: %w", err)
-	}
-	return countDomains(u, domain)
-}
+func GetDomainStatNew(r io.Reader, domain string) (DomainStat, error) {
+	domainStat := make(DomainStat)
+	domainStatDot := "." + strings.ToLower(domain)
 
-type users [100_000]User
+	scanner := bufio.NewScanner(r)
+	scanner.Split(bufio.ScanLines)
 
-func getUsers(r io.Reader) (result users, err error) {
-	content, err := ioutil.ReadAll(r)
-	if err != nil {
-		return
-	}
-
-	lines := strings.Split(string(content), "\n")
-	for i, line := range lines {
-		var user User
-		if err = json.Unmarshal([]byte(line), &user); err != nil {
-			return
+	for scanner.Scan() {
+		res := scanner.Text()
+		if !strings.Contains(res, domain) {
+			continue
 		}
-		result[i] = user
-	}
-	return
-}
 
-func countDomains(u users, domain string) (DomainStat, error) {
-	result := make(DomainStat)
-
-	for _, user := range u {
-		matched, err := regexp.Match("\\."+domain, []byte(user.Email))
+		user := &User{}
+		err := easyjson.Unmarshal([]byte(res), user)
 		if err != nil {
 			return nil, err
 		}
 
-		if matched {
-			num := result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]
-			num++
-			result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])] = num
+		domain := strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])
+		if !strings.Contains(domain, domainStatDot) {
+			continue
 		}
+
+		domainStat[domain]++
 	}
-	return result, nil
+
+	return domainStat, nil
 }
+
+func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
+	return GetDomainStatNew(r, domain)
+	// u, err := getUsers(r)
+	// if err != nil {
+	//	return nil, fmt.Errorf("get users error: %w", err)
+	// }
+	// return countDomains(u, domain)
+}
+
+// type users [100_000]User
+//
+// func getUsers(r io.Reader) (result users, err error) {
+//	content, err := ioutil.ReadAll(r)
+//	if err != nil {
+//		return
+//	}
+//
+//	lines := strings.Split(string(content), "\n")
+//	for i, line := range lines {
+//		var user User
+//		if err = json.Unmarshal([]byte(line), &user); err != nil {
+//			return
+//		}
+//		result[i] = user
+//	}
+//	return
+//}
+//
+// func countDomains(u users, domain string) (DomainStat, error) {
+//	result := make(DomainStat)
+//
+//	for _, user := range u {
+//		matched, err := regexp.Match("\\."+domain, []byte(user.Email))
+//		if err != nil {
+//			return nil, err
+//		}
+//
+//		if matched {
+//			num := result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]
+//			num++
+//			result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])] = num
+//		}
+//	}
+//	return result, nil
+//}