--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -4,8 +4,6 @@ import (
 	"bufio"
 	"io"
 	"strings"
-
-	"github.com/mailru/easyjson"
 )
 
 type User struct {
@@ -33,18 +31,19 @@ func GetDomainStatNew(r io.Reader, domain string) (DomainStat, error) {
 			continue
 		}
 
-		user := &User{}
-		err := easyjson.Unmarshal([]byte(res), user)
-		if err != nil {
+		//user := &User{}
+		var user User
+
+		if err := user.UnmarshalJSON([]byte(res)); err != nil {
 			return nil, err
 		}
 
-		domain := strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])
-		if !strings.Contains(domain, domainStatDot) {
+		domainRes := strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])
+		if !strings.Contains(domainRes, domainStatDot) {
 			continue
 		}
 
-		domainStat[domain]++
+		domainStat[domainRes]++
 	}
 
 	return domainStat, nil
@@ -52,46 +51,4 @@ func GetDomainStatNew(r io.Reader, domain string) (DomainStat, error) {
 
 func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
 	return GetDomainStatNew(r, domain)
-	// u, err := getUsers(r)
-	// if err != nil {
-	//	return nil, fmt.Errorf("get users error: %w", err)
-	// }
-	// return countDomains(u, domain)
 }
-
-// type users [100_000]User
-//
-// func getUsers(r io.Reader) (result users, err error) {
-//	content, err := ioutil.ReadAll(r)
-//	if err != nil {
-//		return
-//	}
-//
-//	lines := strings.Split(string(content), "\n")
-//	for i, line := range lines {
-//		var user User
-//		if err = json.Unmarshal([]byte(line), &user); err != nil {
-//			return
-//		}
-//		result[i] = user
-//	}
-//	return
-//}
-//
-// func countDomains(u users, domain string) (DomainStat, error) {
-//	result := make(DomainStat)
-//
-//	for _, user := range u {
-//		matched, err := regexp.Match("\\."+domain, []byte(user.Email))
-//		if err != nil {
-//			return nil, err
-//		}
-//
-//		if matched {
-//			num := result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]
-//			num++
-//			result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])] = num
-//		}
-//	}
-//	return result, nil
-//}